Integrate board resource into order resource
