+++ hexa_delivery/lib/bloc/board_bloc.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:hexa_delivery/model/category.dart';
 import 'package:hexa_delivery/model/dto.dart';
-import 'package:hexa_delivery/resources/board_order_provider.dart';
+import 'package:hexa_delivery/resources/order_resource.dart';
 
 class BoardBloc {
   final StreamController<List<OrderDescDTO>> _orderStream = StreamController();
@@ -12,7 +12,7 @@ class BoardBloc {
   Stream<List<OrderDescDTO>> get getOrderStream => _orderStream.stream;
 
   void fetchNextPage({Category? category, int? uid}) async {
-    List<OrderDescDTO> fetchList = await BoardResource.getOrders(category: category, uid: uid, pageIndex: _page);
+    List<OrderDescDTO> fetchList = await OrderResource.getOrders(category: category, uid: uid, pageIndex: _page);
     _orderList.addAll(fetchList);
     _orderStream.sink.add(_orderList);
     if (fetchList.isNotEmpty) _page += 1;
